MR_python/requirements.txt
python=2.7.13
numpy==1.13.1
pandas==0.20.3
scipy==0.19.1
google-api-python-client
google-compute-engine
google-cloud-storage
sh
retrying
pyyaml
oauth2client==3.0.0
